+++ tests/Database/SelectFields/ValidateFieldTests/PrivacyArgs.php
@@ -3,6 +3,7 @@
 declare(strict_types = 1);
 namespace Rebing\GraphQL\Tests\Database\SelectFields\ValidateFieldTests;
 
+use GraphQL\Type\Definition\ResolveInfo;
 use PHPUnit\Framework\Assert;
 use Rebing\GraphQL\Support\Privacy;
 
@@ -11,7 +12,7 @@ class PrivacyArgs extends Privacy
     /**
      * @inheritDoc
      */
-    public function validate(array $fieldArgs, $queryContext = null): bool
+    public function validate(mixed $root, array $fieldArgs, mixed $queryContext = null, ?ResolveInfo $resolveInfo = null): bool
     {
         $expectedQueryArgs = [
             'arg_from_field' => true,