+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 [Next release](https://github.com/rebing/graphql-laravel/compare/9.17.0...master)
 
 ## Breaking changes
+- `Privacy::validate()` and closure signature changed [\#1251 / mfn](https://github.com/rebing/graphql-laravel/pull/1251)
+  new `mixed $root` first parameter, new optional `?ResolveInfo $resolveInfo` fourth parameter, `$queryContext` now typed as `mixed`
 - Remove `$getSelectFields` parameter from `Field::authorize()` [\#1250 / mfn](https://github.com/rebing/graphql-laravel/pull/1250)
   it has been non-functional since half a decade